Rename gearApi to gearService
