+++ website/test-balance/src/gear/index.ts
@@ -5,9 +5,9 @@ import { initLogger } from '@gear-js/common';
 
 import config from '../config/configuration';
 import { ResponseTransferBalance } from './types';
-import { transferService } from '../domain/transfer/transfer.service';
+import { transferService } from '../services/transfer/transfer.service';
 
-let gear: GearApi;
+let gearApi: GearApi;
 let accountGR: KeyringPair;
 let rootAccountGR: KeyringPair;
 let accountBalanceGR: BN;
@@ -16,11 +16,11 @@ let balanceToTransferGR: BN;
 const logger = initLogger('TEST_BALANCE_GEAR');
 
 async function connect() {
-  gear = await GearApi.create({ providerAddress: config.gear.providerAddress });
-  gear.on('error', () => {
+  gearApi = await GearApi.create({ providerAddress: config.gear.providerAddress });
+  gearApi.on('error', () => {
     GearApi.create({ providerAddress: config.gear.providerAddress }).then(
       (newApi) => {
-        gear = newApi;
+        gearApi = newApi;
       },
       (error) => {
         logger.error(`Could not reconnect error: ${error}`);
@@ -28,7 +28,7 @@ async function connect() {
       },
     );
   });
-  logger.info(`Connected to ${await gear.chain()} with genesis ${getGenesisHash()}`);
+  logger.info(`Connected to ${await gearApi.chain()} with genesis ${getGenesisHash()}`);
 
   const [account, rootAccountByGearKeyring] = await Promise.all([
     GearKeyring.fromSeed(config.gear.accountSeed),
@@ -47,8 +47,8 @@ async function connect() {
 
 async function transferBalance(
   to: string,
-  from: KeyringPair = this.account,
-  balance: BN = this.accountBalance,
+  from: KeyringPair = accountGR,
+  balance: BN = accountBalanceGR,
 ): Promise<ResponseTransferBalance> {
   try {
     await transfer(from, to, balance);
@@ -56,8 +56,8 @@ async function transferBalance(
     logger.error(error);
     return { error: `Transfer balance from ${from} to ${to} failed` };
   }
-  if (to !== this.account.address) {
-    await transferService.setTransferDate(to, this.getGenesisHash);
+  if (to !== accountGR.address) {
+    await transferService.setTransferDate(to, getGenesisHash());
   }
   return { status: 'ok', transferredBalance: balance.toString() };
 }
@@ -67,9 +67,9 @@ async function transfer(
   to: string,
   balance: BN = accountBalanceGR,
 ): Promise<TransferData> {
-  gear.balance.transfer(to, balance);
+  gearApi.balance.transfer(to, balance);
   return new Promise((resolve, reject) => {
-    gear.balance.signAndSend(from, ({ events }) => {
+    gearApi.balance.signAndSend(from, ({ events }) => {
       events.forEach(({ event: { method, data } }) => {
         if (method === 'Transfer') {
           resolve(data as TransferData);
@@ -91,7 +91,7 @@ async function setRootAccountSeedByGearKeyring(): Promise<KeyringPair> {
 }
 
 async function isSmallAccountBalance(): Promise<boolean> {
-  const balance = await gear.balance.findOut(accountGR.address);
+  const balance = await gearApi.balance.findOut(accountGR.address);
   if (balance.lt(accountBalanceGR)) {
     return true;
   }
@@ -99,7 +99,7 @@ async function isSmallAccountBalance(): Promise<boolean> {
 }
 
 function getGenesisHash(): string {
-  return gear.genesisHash.toHex();
+  return gearApi.genesisHash.toHex();
 }
 
-export const gearApi = { connect, getGenesisHash, transferBalance };
+export const gearService = { connect, getGenesisHash, transferBalance };
